+++ CHANGELOG.md
@@ -18,9 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Inheritance of method docstrings with multiple inheritance such that at least
+- Inheritance of method docstring with multiple inheritance such that at least
   the first class is not set to inherit.
 - Performance improvements.
+- Detecting missing arguments when there was no inheritance for a docstring.
+- Inheriting from `object.__init__` when the first class in the hierarchy
+  has no docstring for `__init__`.
 
 ### Removed